requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
django-environ>=0.11

# παραγωγή
whitenoise>=6.7

# αν χρησιμοποιείς Supabase/Postgres
psycopg2-binary>=2.9

# Cloudinary (χρησιμοποιείς CloudinaryField)
cloudinary>=1.39
